+++ src/pyramid_bimt/models/group.py
@@ -194,7 +194,7 @@ class Group(Base, BaseMixin):
         return Group.query.filter_by(product_id=product_id).first()
 
     @classmethod
-    def get_all(class_, order_by='name', filter_by=None, limit=1000):
+    def get_all(class_, order_by='name', filter_by=None, limit=None):
         """Return all groups.
 
         filter_by: dict -> {'name': 'foo'}
@@ -206,5 +206,6 @@ class Group(Base, BaseMixin):
         q = q.order_by(getattr(Group, order_by))
         if filter_by:
             q = q.filter_by(**filter_by)
-        q = q.limit(limit)
+        if limit:
+            q = q.limit(limit)
         return q